refactor: move arrow to arrow reader
