+++ src/arrow_reader/column/boolean.rs
@@ -1,9 +1,9 @@
 use snafu::OptionExt;
 
+use crate::arrow_reader::column::present::new_present_iter;
+use crate::arrow_reader::column::{Column, NullableIterator};
 use crate::error::{InvalidColumnSnafu, Result};
 use crate::proto::stream::Kind;
-use crate::reader::column::present::new_present_iter;
-use crate::reader::column::{Column, NullableIterator};
 use crate::reader::decode::boolean_rle::BooleanIter;
 
 pub fn new_boolean_iter(column: &Column) -> Result<NullableIterator<bool>> {
