+++ src/arrow_reader/column/date.rs
@@ -1,10 +1,10 @@
 use chrono::{Days, NaiveDate};
 use snafu::OptionExt;
 
+use crate::arrow_reader::column::present::new_present_iter;
+use crate::arrow_reader::column::{Column, NullableIterator};
 use crate::error::{self, Result};
 use crate::proto::stream::Kind;
-use crate::reader::column::present::new_present_iter;
-use crate::reader::column::{Column, NullableIterator};
 use crate::reader::decode::rle_v2::RleReaderV2;
 
 pub const UNIX_EPOCH_FROM_CE: i32 = 719_163;
